requirements.txt
PyYAML
requests
lxml
beautifulsoup4
boto3
